+++ CHANGELOG.md
@@ -9,10 +9,12 @@ This file is used to list changes made in each version of the keepalived cookboo
   - Removed property `config_name`
   - Removed property `content`, this is now build up from the supplied properties
   - Removed property `path`
+  - Removed property `exists`
   - Added property `conf_directory`, defaulted to: `/etc/keepalived/conf.d`
   - Added property `config_file`, defaulted to: `::File.join(conf_directory, 'global_defs.conf')`
   - Added property `cookbook`, defaulted to: `keepalived`
   - Added property `source`, defaulted to `global_defs.conf.erb`
+  - Added property `extra_options`
 
 ## 3.1.1 (2018-01-10)